refactor: extract product cvss metrics aggregation
