+++ csaf-rs/src/csaf/aggregation/mod.rs
@@ -1,1 +1,2 @@
+pub mod product_cvss_metrics;
 pub mod product_status;
