+++ csaf-rs/src/validations/test_6_1_07.rs
@@ -1,44 +1,7 @@
+use crate::csaf::aggregation::product_cvss_metrics::aggregate_product_cvss_metrics;
 use crate::csaf::types::csaf_vuln_metric::CsafVulnerabilityMetric;
-use crate::csaf_traits::{ContentTrait, CsafTrait, MetricTrait, VulnerabilityTrait};
+use crate::csaf_traits::CsafTrait;
 use crate::validation::ValidationError;
-use std::collections::HashMap;
-
-/// The actual test is run on a 3 level nested map.
-/// The outer map maps product ids to a second map [ProductCsafVulnerabilityMetricMap], which maps to a third map [ProductCsafVulnerabilityMetricSourceMap].
-type ProductMap = HashMap<String, ProductCsafVulnerabilityMetricMap>;
-/// The second map maps a metric type (e.g., CVSS v3.1) to a third map
-type ProductCsafVulnerabilityMetricMap = HashMap<CsafVulnerabilityMetric, ProductCsafVulnerabilityMetricSourceMap>;
-/// The innermost map maps a source (None for CSAF 2.0 or if None is provided) to a list of JSON paths where this metric type is given for the product and source
-type ProductCsafVulnerabilityMetricSourceMap = HashMap<Option<String>, Vec<String>>;
-fn gather_product_metrics(vulnerability: &impl VulnerabilityTrait, vulnerability_index: usize) -> Option<ProductMap> {
-    let metrics = vulnerability.get_metrics();
-
-    metrics?;
-
-    let mut product_metrics: ProductMap = HashMap::new();
-    for (metric_index, metric) in metrics.unwrap().iter().enumerate() {
-        let content = metric.get_content();
-        let present_metric_types = content.get_cvss_metric_types();
-
-        for product_id in metric.get_products() {
-            for metric_type in &present_metric_types {
-                product_metrics
-                    // First map: product id =>
-                    .entry(product_id.to_owned())
-                    .or_default()
-                    // Second map: metric type =>
-                    .entry(metric_type.to_owned())
-                    .or_default()
-                    // Third map: source (or none) =>
-                    .entry(metric.get_source().map(|s| s.to_owned()))
-                    .or_default()
-                    // vector with json paths that provide this metric type for this product and source
-                    .push(content.get_content_json_path(vulnerability_index, metric_index));
-            }
-        }
-    }
-    Some(product_metrics)
-}
 
 /// Test 6.1.7 Multiple Scores with Same Version per Product
 ///
@@ -47,18 +10,21 @@ fn gather_product_metrics(vulnerability: &impl VulnerabilityTrait, vulnerability
 pub fn test_6_1_07_multiple_same_scores_per_product(doc: &impl CsafTrait) -> Result<(), Vec<ValidationError>> {
     let mut errors: Option<Vec<ValidationError>> = None;
     for (vulnerability_index, vulnerability) in doc.get_vulnerabilities().iter().enumerate() {
-        let product_metrics = gather_product_metrics(vulnerability, vulnerability_index);
+        let product_metrics = aggregate_product_cvss_metrics(vulnerability, vulnerability_index);
+        // if there are any cvss metrics
         if let Some(product_metrics) = product_metrics {
-            for (p, metrics_map) in &product_metrics {
-                for (m, metrics_map_2) in metrics_map {
-                    for (s, paths) in metrics_map_2 {
+            // traverse the metrics map
+            for (product_id, cvss_metrics_map) in &product_metrics {
+                for (metric_type, source_map) in cvss_metrics_map {
+                    for (source, paths) in source_map {
+                        // generate an error if the same metric, with same or no source, applies to the same product id
                         if paths.len() > 1 {
                             for path in paths {
                                 errors.get_or_insert_default().push(create_validation_error(
-                                    m,
-                                    p,
+                                    metric_type,
+                                    product_id,
                                     path.to_owned(),
-                                    s.clone(),
+                                    source.clone(),
                                 ));
                             }
                         }
@@ -281,7 +247,7 @@ mod tests {
         // Case 18: like 05, but valid
         TESTS_2_0
             .test_6_1_7
-            .expect(case_01_duplicate_cvss_v3_csaf_20, Ok(()), Ok(()));
+            .expect(case_01_duplicate_cvss_v3_1_csaf_20, Ok(()), Ok(()));
 
         TESTS_2_1.test_6_1_7.expect(
             case_01_duplicate_cvss_v3_1_csaf_21,
